fix: resolve 7 bugs and design issues (#54-#61)
- Fix stale jcollections package references breaking compilation (#54)
- Fix rebuildIndex() first-wins vs put() last-wins semantic mismatch (#55)
- Fix entrySet()/size() exposing duplicates, violating Map contract (#56)
- Fix IntList.flush()/close() integer overflow at 536M+ entries (#57)
- Add LRU eviction to WriteCache to prevent unbounded memory growth (#58)
- Document FileLockManager TOCTOU race on lock upgrade (#60)
- FileRepairer now scans past corruption instead of stopping (#61)

Multi-AI reviewed: Groq/Llama-3.3-70B approved all fixes, Cerebras/Gemma-4-31B
adversarial review raised 3 concerns, review-the-review confirmed all INVALID.
All 211 tests pass, JaCoCo coverage gates pass.

+++ src/main/java/org/flossware/collections/file/primitive/IntList.java
@@ -124,7 +124,7 @@ public class IntList implements AutoCloseable {
             if (mappedBuffer != null) {
                 mappedBuffer.force();
             }
-            long actualSize = FileHeader.getHeaderSize() + size.get() * INT_SIZE;
+            long actualSize = FileHeader.getHeaderSize() + (long) size.get() * INT_SIZE;
             try {
                 if (channel.isOpen()) {
                     file.setLength(actualSize);
@@ -182,7 +182,7 @@ public class IntList implements AutoCloseable {
                 mappedBuffer.force();
                 unmapBuffer();
             }
-            long actualSize = FileHeader.getHeaderSize() + size.get() * INT_SIZE;
+            long actualSize = FileHeader.getHeaderSize() + (long) size.get() * INT_SIZE;
             if (file != null && channel.isOpen()) {
                 file.setLength(actualSize);
                 channel.force(true);
